Section 1: Game Setup
Created a new 3D project in Godot with all assets imported that I will be using in the game. So far, we have 2 Scenes, the level and the player, and one Script for controlling very basic player movement. Very barebones right now, but the next commit will focus more on player movement and camera controls.