Section 3: Animating the Player
To start, I created a character rig scene and added it to the Player with one of the Knight meshes exposed. I then added a simple animation tree to control the blending of character movement and attacking (slashing). Added a bone attachment to place the sword and shield in the correct hands of the player. Finally, I added a ray cast node to the Player's hand holding the sword (right hand) to toggle collision while swinging the sword. Now we need something to attack. Next commit will focus on creating basic enemies that will deal damage to and receive damage from the Player.