Section 4: Enemy Character
Created a scene for the Enemy character with animations for heavy attack when the Player is in range of said Enemy. Also added this heavy attack to the Player's animation tree, which is triggered by the right mouse button. Added a health component used by all characters to track damage points. Added animations for all characters when defeated (health component is zero). Finally, I added a Dash ability to the character, triggered by the shift key, complete with visual effects to sell the dash animation. The next commit will focus on RPG statistics to better track health, damage, etc.