Section 6: User Interface
Started by importing some assets used in the UI. Added a UI that is displayed in the level scene that shows the player's health, xp (experience points), and level number. Added a script to the newly created User Interface scene that updates the progress bars for these attributes. Created an Inventory scene (accessed by the I or TAB key) that shows the RPG stats in a clear format. These attributes are updated when the player levels up or gains xp. Finally, I added a character preview to the Inventory screen, allowing players to preview their inventory items. This feature will be implemented in a future commit.