Section 7: Enemy Navigation
I added a NavigationRegion3D node to the Enemy scene to allow the Enemy to move around the Level scene. Coded in Navigational Agents so the Enemy can now "see" the Player and give chase while avoiding the path of the other Enemies. Finally, I set up a function to randomly assign equipment (weapon and shield) to each Enemy spawned at the start of the game. This section was pretty straightforward with minimal coding. The next commit will focus on the graphics, look, and feel of the game.