Section: 8: Graphics and Aesthetics
Did some touch-ups on the light and graphics, and also imported a few more features into the level scene to make it more realistic. As the files are getting more complex, I tagged all the scene files in the Git LFS system for better version control storage.  We will get back to coding in the next commit, where I will set up the actual Inventory system in the user interface.